requirements.txt
falcon==2.0.0
hvac==0.9.6
Werkzeug==0.16.0
msgpack-python==0.5.6
pytest==5.3.1
coverage==4.5.4
SQLAlchemy==1.3.11
SQLAlchemy-Utils==0.35.0
mixer==6.1.3
marshmallow==3.2.2
marshmallow-sqlalchemy==0.19.0
PyMySQL[rsa]==0.9.3
